requirements.txt
torch>=2.1
torchvision>=0.16
timm>=0.9.16
transformers>=4.42.0
tokenizers>=0.15.2
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
tqdm>=4.66.0
Pillow>=10.0.0
